analysispipeline/requirements.txt
flask
python-dotenv
flask-mysqldb
groq

